requirements.txt
flask
pymongo
requests
env
python-dotenv
eventlet
httpx
bloompy
pyinterval
scikit-learn
pyyaml
pandas
flask_cors
